Task2-Array. Add some tests and divided classes in different files.
Add tests for all constructors of Vector and push_back and pop_back
methods of Vector. Place all classes and realizations of them in
different files.
